requirements.txt
Pillow
numpy
opencv-python

mediapipe

diffusers 
transformers 
accelerate 
safetensors 
